fix(ionic app scripts version): updated to solve this error (build dev failed: Cannot set property 'fileSystem' of null)
