--- app/pages/details/details.scss
@@ -1,3 +1,0 @@
-.details {
-
-}
