--- app/pages/home/home.scss
@@ -1,3 +1,0 @@
-.home {
-  
-}
