+++ ionic.config.json
@@ -1,6 +1,6 @@
 {
-  "name": "ionic2-app-base",
+  "name": "ionic2-tutorial-pouchdb",
   "app_id": "",
-  "typescript": true,
-  "v2": true
+  "v2": true,
+  "typescript": true
 }
