+++ package.json
@@ -1,29 +1,49 @@
 {
+  "name": "ionic-hello-world",
+  "author": "Ionic Framework",
+  "homepage": "http://ionicframework.com/",
+  "private": true,
+  "scripts": {
+    "ionic:build": "ionic-app-scripts build",
+    "ionic:serve": "ionic-app-scripts serve"
+  },
   "dependencies": {
-    "@angular/common": "2.0.0-rc.3",
-    "@angular/compiler": "2.0.0-rc.3",
-    "@angular/core": "2.0.0-rc.3",
-    "@angular/http": "2.0.0-rc.3",
-    "@angular/platform-browser": "2.0.0-rc.3",
-    "@angular/platform-browser-dynamic": "2.0.0-rc.3",
-    "es6-shim": "^0.35.0",
-    "ionic-angular": "2.0.0-beta.10",
-    "ionic-native": "1.2.4",
+    "@angular/common": "2.2.1",
+    "@angular/compiler": "2.2.1",
+    "@angular/compiler-cli": "2.2.1",
+    "@angular/core": "2.2.1",
+    "@angular/forms": "2.2.1",
+    "@angular/http": "2.2.1",
+    "@angular/platform-browser": "2.2.1",
+    "@angular/platform-browser-dynamic": "2.2.1",
+    "@angular/platform-server": "2.2.1",
+    "@ionic/storage": "1.1.6",
+    "ionic-angular": "2.0.0-rc.4",
+    "ionic-native": "2.2.11",
     "ionicons": "3.0.0",
-    "pouchdb": "^5.4.4",
-    "reflect-metadata": "^0.1.3",
-    "rxjs": "5.0.0-beta.6",
-    "zone.js": "^0.6.12"
+    "pouchdb": "^6.1.0",
+    "rxjs": "5.0.0-beta.12",
+    "zone.js": "0.6.26"
   },
   "devDependencies": {
-    "del": "2.2.0",
-    "gulp": "3.9.1",
-    "gulp-watch": "4.3.5",
-    "ionic-gulp-browserify-typescript": "2.0.0",
-    "ionic-gulp-fonts-copy": "^1.0.0",
-    "ionic-gulp-html-copy": "^1.0.0",
-    "ionic-gulp-sass-build": "^1.0.0",
-    "ionic-gulp-scripts-copy": "^2.0.0",
-    "run-sequence": "1.1.5"
-  }
+    "@ionic/app-scripts": "1.0.0",
+    "typescript": "2.0.9"
+  },
+  "cordovaPlugins": [
+    "cordova-plugin-whitelist",
+    "cordova-plugin-console",
+    "cordova-plugin-statusbar",
+    "cordova-plugin-device",
+    "cordova-plugin-splashscreen",
+    "ionic-plugin-keyboard"
+  ],
+  "cordovaPlatforms": [
+    "ios",
+    {
+      "platform": "ios",
+      "version": "",
+      "locator": "ios"
+    }
+  ],
+  "description": "ionic2-tutorial-pouchdb: An Ionic project"
 }
