+++ src/app/app.html
@@ -1,0 +1,3 @@
+<script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=Intl.~locale.en"></script>
+
+<ion-nav [root]="rootPage"></ion-nav>
