+++ src/pages/details/details.scss
@@ -1,0 +1,3 @@
+page-details {
+
+}
