+++ src/pages/home/home.html
@@ -1,0 +1,21 @@
+<ion-header>
+  <ion-navbar>
+    <ion-title>
+      🎂  Birthdays  🎉
+    </ion-title>
+    <ion-buttons end>
+      <button ion-button (click)="showDetail()">
+          <ion-icon name="add"></ion-icon>
+      </button>
+    </ion-buttons>
+  </ion-navbar>
+</ion-header>
+
+<ion-content>
+    <ion-list inset>
+        <ion-item *ngFor="let birthday of birthdays" (click)="showDetail(birthday)">
+          <div item-left>{{ birthday.Name }}</div>
+          <div item-right>{{ birthday.Date | date:'yMMMMd' }}</div>
+        </ion-item>
+      </ion-list>
+</ion-content>
