+++ src/pages/home/home.scss
@@ -1,0 +1,3 @@
+page-home {
+
+}
