--- typings/index.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="globals/es6-shim/index.d.ts" />
-/// <reference path="globals/require/index.d.ts" />
